sud_loyihasi/requirements.txt
# AI-Audit — Sud Qarorlari Korrupsion Risk Tahlil Tizimi
# Python 3.10+ tavsiya etiladi

# ============================================================
# ASOSIY (kerakli)
# ============================================================
requests>=2.31.0          # HTTP so'rovlar (parser)
openai>=1.10.0            # DeepSeek API (OpenAI-compatible client)
pdfplumber>=0.10.0        # PDF dan matn ajratish
python-docx>=1.1.0        # DOCX dan matn ajratish
olefile>=0.46             # Eski .doc fayllar uchun fallback


# ============================================================
# WINDOWS QO'SHIMCHA (faqat Windows + MS Word bo'lsa)
# ============================================================
# pywin32>=306            # MS Word COM orqali .doc ochish
                          # Qo'lda o'rnating: pip install pywin32


# ============================================================
# OPTIONAL — TAHLIL VA EKSPORT
# ============================================================
pandas>=2.0.0             # Statistika va data manipulation
numpy>=1.24.0             # Numerical hisoblar
openpyxl>=3.1.0           # Excel export


# ============================================================
# OPTIONAL — WEB APP (agar Flask app ishlatsangiz)
# ============================================================
flask>=3.0.0              # Web framework
werkzeug>=3.0.0           # Flask dependency


# ============================================================
# DEVELOPMENT (ixtiyoriy)
# ============================================================
# pytest>=7.4.0           # Testlash
# black>=23.0.0           # Kod formatlash
